+++ sdk/cargo-build-sbf/tests/crates.rs
@@ -176,3 +176,17 @@ fn test_sbfv2() {
         .success();
     clean_target("noop");
 }
+
+#[test]
+#[serial]
+fn test_package_metadata_tools_version() {
+    run_cargo_build("package-metadata", &[], false);
+    clean_target("package-metadata");
+}
+
+#[test]
+#[serial]
+fn test_workspace_metadata_tools_version() {
+    run_cargo_build("workspace-metadata", &[], false);
+    clean_target("workspace-metadata");
+}